added auth to task-delete

+++ apps/tasks-delete/src/app/task/task.controller.ts
@@ -1,6 +1,5 @@
-import { Controller, Get, Post, Body, Param, Put, Delete } from '@nestjs/common';
+import { Controller, Param, Headers, Delete, ForbiddenException, UnauthorizedException } from '@nestjs/common';
 import { TaskService } from './task.service';
-import { Task } from './task.entity';
 
 @Controller('tasks')
 export class TaskController {
@@ -19,7 +18,20 @@ export class TaskController {
 //   }
 
   @Delete(':id')
-  deleteTask(@Param('id') id: number): Promise<void> {
+  async deleteTask(@Param('id') id: number,@Headers('Authorization') authHeader: string): Promise<void> {
+
+    if (!authHeader) {
+      throw new UnauthorizedException('Token is required');
+    }
+
+    const isValidate = await this.tasksService.validateToken(authHeader);
+
+    console.log('isValidate', authHeader);
+    if(!isValidate ){
+
+      throw new ForbiddenException('Invalid token!!!');
+    }
+
     return this.tasksService.deleteTask(id);
   }
 }
